requirements.txt
dvc==2.9.3
numpy==1.22.0
pandas==1.3.5
scikit-learn==1.0.2
matplotlib==3.5.1
pyyaml==6.0